task/sql: task_id is no longer a foreign key.
When we delete a task.Task, we remove it from the database. Therefore,
if we try to create a task.Event (such as one with type
task.EventTypeDeleted), we fail since no task.Task exists with that ID.
This seems super wrong, so if we find out a better way to do this in the
future then let's change this.